Create AdminBackupBuddy_SelectedAppData.bat
This is a modification to the original file. This only looks for and backs up certain selected AppData. It skips the Outlook .ost file, (assuming it can just be re-created from mail server.) It specifically looks for and copies any .pst files. It specifically backs up AppData for applications such as Spotify, Google, Mozilla, Adobe, Opera, and some Microsoft app settings.